+++ avian2d/src/lib.rs
@@ -260,14 +260,17 @@ fn update_proximity_sensors_system(
                     cast_origin.truncate().adjust_precision(),
                     0.0,
                     cast_direction_2d,
-                    sensor.cast_range,
-                    true,
-                    query_filter,
+                    &ShapeCastConfig {
+                        max_distance: sensor.cast_range,
+                        ignore_origin_penetration: true,
+                        ..default()
+                    },
+                    &query_filter,
                     #[allow(clippy::useless_conversion)]
                     |shape_hit_data| {
                         apply_cast(CastResult {
                             entity: shape_hit_data.entity,
-                            proximity: shape_hit_data.time_of_impact,
+                            proximity: shape_hit_data.distance,
                             intersection_point: shape_hit_data.point1,
                             normal: Dir3::new(shape_hit_data.normal1.extend(0.0).f32())
                                 .unwrap_or_else(|_| -cast_direction),
@@ -280,13 +283,13 @@ fn update_proximity_sensors_system(
                     cast_direction_2d,
                     sensor.cast_range,
                     true,
-                    query_filter,
+                    &query_filter,
                     |ray_hit_data| {
                         apply_cast(CastResult {
                             entity: ray_hit_data.entity,
-                            proximity: ray_hit_data.time_of_impact,
+                            proximity: ray_hit_data.distance,
                             intersection_point: cast_origin.truncate().adjust_precision()
-                                + ray_hit_data.time_of_impact.adjust_precision()
+                                + ray_hit_data.distance.adjust_precision()
                                     * cast_direction_2d.adjust_precision(),
                             normal: Dir3::new(ray_hit_data.normal.extend(0.0).f32())
                                 .unwrap_or_else(|_| -cast_direction),
@@ -317,7 +320,7 @@ fn update_obstacle_radars_system(
             &Collider::rectangle(2.0 * radar.radius, radar.height),
             radar_position.0,
             Default::default(),
-            Default::default(),
+            &SpatialQueryFilter::DEFAULT,
             |obstacle_entity| {
                 if radar_owner_entity == obstacle_entity {
                     return true;
@@ -335,8 +338,8 @@ fn apply_motors_system(
         &TnuaMotor,
         &mut LinearVelocity,
         &mut AngularVelocity,
-        &Mass,
-        &Inertia,
+        &ComputedMass,
+        &ComputedAngularInertia,
         &mut ExternalForce,
         &mut ExternalTorque,
         Option<&TnuaToggle>,
@@ -364,7 +367,7 @@ fn apply_motors_system(
             linare_velocity.0 += motor.lin.boost.truncate();
         }
         if motor.lin.acceleration.is_finite() {
-            external_force.set_force(motor.lin.acceleration.truncate() * mass.0);
+            external_force.set_force(motor.lin.acceleration.truncate() * mass.value());
         }
         if motor.ang.boost.is_finite() {
             angular_velocity.0 += motor.ang.boost.z;
@@ -373,7 +376,7 @@ fn apply_motors_system(
             external_torque.set_torque(
                 // NOTE: I did not actually verify that this is the correct formula. Nothing uses
                 // angular acceleration yet - only angular impulses.
-                inertia.0 * motor.ang.acceleration.z,
+                inertia.value() * motor.ang.acceleration.z,
             );
         }
     }