+++ CHANGELOG.md
@@ -17,6 +17,14 @@ NOTE: Subcrates have their own changelogs: [bevy-tnua-physics-integration-layer]
 - `vertical_displacement` and `force_forward` fields for `TnuaBuiltinJump`.
   This is useful for wall jumps.
 
+## 0.21.0 - 2024-12-13
+### Changed
+- Upgrade to Bevy 0.15.
+
+### Removed
+- `TnuaControllerBundle`. It is no longer needed since `TnuaController` uses
+  Bevy 0.15's required components feature.
+
 ## 0.20.0 - 2024-10-12
 ### Added
 - A `TnuaBuiltinKnockback` action for applying knockback that will not be